fix dtype mismatch

+++ cs336_systems/flash_attention/flash_attention_triton.py
@@ -227,8 +227,6 @@ def flash_bwd_kernel(
         order=(1, 0),
     )
 
-    scale = scale.to(tl.float16)
-    
     Qi = tl.load(Q_block_ptr, boundary_check=(0, 1), padding_option="zero").to(tl.float16)
     Oi = tl.load(O_block_ptr, boundary_check=(0, 1), padding_option="zero").to(tl.float16)
     dOi = tl.load(dO_block_ptr, boundary_check=(0, 1), padding_option="zero").to(tl.float16)
@@ -240,14 +238,14 @@ def flash_bwd_kernel(
         Kj = tl.load(K_block_ptr, boundary_check=(0, 1), padding_option="zero").to(tl.float16)
         Vj = tl.load(V_block_ptr, boundary_check=(0, 1), padding_option="zero").to(tl.float16)
 
-        # (Bq, D) float16 * (D, Bk) float16 = (Bq, Bk) float16
+        # (Bq, D) float16 * (D, Bk) float16 * scale float32 = (Bq, Bk) float32
         Sij = tl.dot(Qi, tl.trans(Kj)) * scale
 
-        # (Bq, Bk) float16
+        # (Bq, Bk) float32
         Pij = tl.exp(Sij - tl.broadcast_to(Li[:, None], Sij.shape))
 
         # (Bk, Bq) float16 * (Bq, D) float16 = (Bk, D) float16
-        dVi = tl.dot(tl.trans(Pij), dOi)
+        dVi = tl.dot(tl.trans(Pij).to(dOi.dtype), dOi)
         tl.atomic_add(dV_block_ptr, dVi)
 
         # (Bq, D) float16 * (D, Bk) float16 = (Bq, Bk) float16
@@ -255,13 +253,13 @@ def flash_bwd_kernel(
 
         # (Bq,) float16
         Di = tl.sum(Oi * dOi, axis=1)
-        # (Bq, Bk) float16
+        # (Bq, Bk) float32
         dSij = Pij * (dPij - tl.broadcast_to(Di[:, None], dPij.shape))
 
-        # (Bq, Bk) float16 * (Bk, D) float16 = (Bq, D) float16
-        dQi = tl.dot(dSij, Kj, acc=dQi) * scale
-        # (Bk, Bq) float16 * (Bq, D) float16 = (Bk, D) float16
-        dKi = tl.dot(tl.trans(dSij), Qi) * scale
+        # (Bq, Bk) float16 * (Bk, D) float16 * scale float32 = (Bq, D) float32
+        dQi = tl.dot(dSij.to(Kj.dtype), Kj, acc=dQi) * scale
+        # (Bk, Bq) float16 * (Bq, D) float16 * scale float32 = (Bk, D) float32
+        dKi = tl.dot(tl.trans(dSij), Qi.to(dSij.dtype)) * scale
         tl.atomic_add(dK_block_ptr, dKi)
 
         K_block_ptr = K_block_ptr.advance((K_TILE_SIZE, 0))
